% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourbar-legend.r
\name{bar.legend}
\alias{bar.legend}
\title{bar.legend()}
\usage{
bar.legend(x = "topright", y = NULL, legd, colv, max.pt = 1,
  revleg = FALSE, leg.title = NA, cex = 1, max.val = NA,
  horiz = FALSE, left.start = TRUE, roundleg = FALSE, title.cex = 1.25,
  bty = "n", ...)
}
\arguments{
\item{x:}{x position or keyword}

\item{y:}{y position if x is defined}

\item{legd:}{legend vector}

\item{colv:}{colour vector with length of legend - 1}

\item{max.pt:}{size of the legend symbols}

\item{max.val:}{optional maximum value observed if not included in the scale}
}
\description{
Modified legend for colourbars
uses a version of legend() modified to work w Hershey Fonts
}
\examples{
bar.legend()
}
\keyword{graphics,}
\keyword{legend}
