% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-color-palettes.r
\name{bathy.colpal}
\alias{bathy.colpal}
\title{Colour palettes for bathymetry}
\usage{
bathy.colpal(n)
}
\arguments{
\item{n}{number of colours in palette}
}
\description{
bathy.colpal() creates a nice colour palette for
representing oceanography data like bathymetry, going from deep blues to turquoise to light greens
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
bathy.colpal()
}
\keyword{colour}
\keyword{graphics,}
\keyword{oceanography,}
\keyword{palettes}
