% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-dev-size.r
\name{check.dev.size}
\alias{check.dev.size}
\title{check.dev.size}
\usage{
check.dev.size(ww, hh, use.prop = FALSE)
}
\arguments{
\item{ww}{desired width in inches}

\item{hh}{desired height in inches}

\item{use.prop}{if set to TRUE and height greater than 7.5, scales width and height to have height = 7.5in and maintain ww:hh proportion}
}
\description{
This function checks the size of your current graphic device
against width (ww) and height (hh) and reopens one of the specified
dimension if needed
}
\examples{
check.dev.size()
}
\keyword{graphics}
