% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-color-palettes.r
\name{chl.colpal}
\alias{chl.colpal}
\title{Colour palettes for chl-a}
\usage{
chl.colpal(n)
}
\arguments{
\item{n}{number of colours in palette}
}
\description{
chl.colpal() creates a nice colour palette for
representing oceanography data like chl-a, going from deep blues to bright greens, golds and pinks
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
chl.colpal()
}
\keyword{colour}
\keyword{graphics,}
\keyword{oceanography,}
\keyword{palettes}
