% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2transp.r
\name{col2transp}
\alias{col2transp}
\title{col2transp}
\usage{
col2transp(col, tlev = 0.5)
}
\arguments{
\item{col}{colour vector to add transparency to}

\item{tlev}{transparency level desired, higher level means more transparent}
}
\description{
This function returns a transparent version of the input colour vector
}
\examples{
col2transp()

}
\keyword{graphics}
