% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.r
\name{count}
\alias{count}
\title{count}
\usage{
count(x)
}
\arguments{
\item{x}{numeric or character vector to count unique values for}
}
\description{
This function returns the number of unique value in a vector
}
\examples{
count()
}
\keyword{statistics}
\keyword{summary}
