% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-color-palettes.r
\name{cpue.colpal}
\alias{cpue.colpal}
\title{Colour palette for CPUE and abundance}
\usage{
cpue.colpal(n = 12)
}
\arguments{
\item{n}{number of colours in palette}
}
\description{
cpue.colpal() defines a custom colour palette going from blue to
yellow to red for CPUE/abundance plots
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
cpue.colpal()
}
\keyword{colour}
\keyword{graphics,}
\keyword{palettes}
