% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-color-palettes.r
\name{jtl.colpal}
\alias{jtl.colpal}
\title{Joe's tight lycra}
\usage{
jtl.colpal(n)
}
\arguments{
\item{n}{number of colours in palette}
}
\description{
jtl.colpal() creates a nice neutral colour palette
going from beiges to dark blues, inspired by
a memorable kayak ride on the lagoon
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
jtl.colpal()
}
\keyword{Canard}
\keyword{Ilot}
\keyword{colour}
\keyword{graphics,}
\keyword{palettes,}
