% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-ltb-2.R
\name{legend.ltb.2}
\alias{legend.ltb.2}
\title{legend.ltb.2()}
\usage{
legend.ltb.2(x, y = NULL, legend, fill = NULL, col = par("col"),
  border = "black", lty, lwd, pch, angle = 45, density = NULL,
  bty = "o", bg = par("bg"), box.lwd = par("lwd"), box.lty = par("lty"),
  box.col = par("fg"), pt.bg = NA, cex = 1, pt.cex = cex,
  pt.lwd = lwd, xjust = 0, yjust = 1, x.intersp = 1, y.intersp = 1,
  adj = c(0, 0.5), text.width = NULL, text.col = par("col"),
  text.font = NULL, merge = do.lines && has.pch, trace = FALSE,
  plot = TRUE, ncol = 1, horiz = FALSE, title = NULL, inset = 0, xpd,
  title.col = text.col, title.adj = 0.5, seg.len = 2, ...)
}
\arguments{
\item{x:}{x position or keyword}

\item{y:}{y position}

\item{legend:}{vector with legend (see legend() for other arguments)}
}
\description{
Defines a (very slightly) modified version of legend()
to allow for Hershey fonts to be used
}
\examples{
legend.ltb.2()
}
\keyword{graphics,}
\keyword{legend}
