% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-ggplot-themes.r
\name{make.panel.theme}
\alias{make.panel.theme}
\title{make.panel.theme returns a clean ggplot theme well suited for making nice panel plots}
\usage{
make.panel.theme(strip.text.background.fill = "slategray")
}
\arguments{
\item{strip.text.background.fill}{colour for the panel background}
}
\description{
make.panel.theme returns a clean ggplot theme well suited for making nice panel plots
}
\examples{
make.panel.theme
}
\keyword{ggplot,}
\keyword{graphics}
\keyword{theme,}
