% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-color-palettes.r
\name{res.colpal}
\alias{res.colpal}
\alias{resid.colpal}
\title{Colour palettes for residuals}
\usage{
res.colpal(n)
}
\arguments{
\item{n}{number of colours in palette}
}
\description{
res.colpal() and resid.colpal() create a colour palette going from
blues to reds via greys that are useful for residual mapping
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
res.colpal()
}
\keyword{colour}
\keyword{graphics,}
\keyword{palettes}
\keyword{residuals,}
