% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-ggplot-themes.r
\name{shift.ylab}
\alias{shift.ylab}
\title{shift.ylab shifts the position of the y-axis labels in a ggplot object}
\usage{
shift.ylab(ypos = 0.5)
}
\arguments{
\item{ypos}{relative shift, default is 0.5}
}
\description{
shift.ylab shifts the position of the y-axis labels in a ggplot object
}
\examples{
shift.ylab
}
\keyword{ggplot,}
\keyword{graphics}
\keyword{theme,}
