% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer-functions.r
\name{start.timer}
\alias{start.timer}
\title{start.timer() initializes a timer object to time code operations}
\usage{
\method{start}{timer}(messg)
}
\arguments{
\item{messg}{an optional message to print out when the timer is started}
}
\description{
start.timer() initializes a timer object to time code operations
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
start.timer()
}
\keyword{timer}
