% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_timer.r
\name{startTimer}
\alias{startTimer}
\title{startTimer() stops the timer and returns the time in seconds since start_timer() was launched}
\usage{
startTimer(messg)
}
\arguments{
\item{messg}{an optional message to print out when the timer is started}
}
\description{
startTimer() stops the timer and returns the time in seconds since start_timer() was launched
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
startTimer()
}
\keyword{timer}
