% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer-functions.r
\name{stop.timer}
\alias{stop.timer}
\title{stop.timer() stops the timer and returns the time in seconds since start.timer() was launched}
\usage{
stop.timer()
}
\description{
stop.timer() stops the timer and returns the time in seconds since start.timer() was launched
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
stop.timer()
}
\keyword{timer}
