% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_timer.r
\name{stopTimer}
\alias{stopTimer}
\title{stopTimer() stops the timer and returns the time in seconds since start_timer() was launched}
\usage{
stopTimer()
}
\description{
stopTimer() stops the timer and returns the time in seconds since start_timer() was launched
}
\author{
Laura Tremblay-Boyer (laura.t.boyer@gmail.com)
stopTimer()
}
\keyword{timer}
