% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doitall4cygwin.r
\name{doitall4cyg}
\alias{doitall4cyg}
\title{doitall4cygwin.r
doitall4cyg() converts a windows-exec doitall to a unix version}
\usage{
doitall4cyg(x)
}
\arguments{
\item{x}{file name for doitall to convert}
}
\description{
doitall4cygwin.r
doitall4cyg() converts a windows-exec doitall to a unix version
}
\author{
Laura Tremblay-Boyer (lauratb@spc.int)
doitall4cyg()
}
\keyword{mfcl}
